feat(data-sources): implement QueryPanelsAsync + StreamPanelsAsync
Shared implementation lives in SqlServerAoiSourceBase and is inherited
unchanged by both HlyaoiSource (v5.0) and MeaoiSource (v4.3.1). The
IS_LAST_INSPECTION filter is gated by Descriptor.Caps so pre-reflow
naturally skips it (its schema doesn't have the column).

- PanelRow DTO now matches the physical schema exactly: PanelId /
  MachineId / etc. as int32, TestTime as double, OperatorId nullable,
  PanelBarCode non-null (NOT NULL varchar(64)).
- PanelCursor / CardCursor / TestedObjectCursor use int32 to match
  the physical Panel_Id column width.
- Keyset paging on (Panel_Numeric_Date, Panel_Id). Server-side TOP
  (@topCount) with pageSize+1 fetch to detect HasMore, then trim
  the extra row and emit NextCursor pointing at the last kept row.
- Optional IN filters (MachineIds / ProductIds / RecipeIds) compose
  into IN (@m0, @m1, ...) with a 500-item cap to keep parameter
  counts sane. Read-only guards, isolation prelude, WITH (NOLOCK)
  and CommandTimeout inherited from the base helper as before.
- Added GetLatestPanelUtcAsync to IAoiSource - simple MAX
  (Panel_Numeric_Date) probe, useful for UI freshness and for
  windowing default queries against each source's own data.
- Smoke test extended: fetches the source's latest panel, then
  queries the 60 days ending there and prints the first 5 rows.
  Verified live against both DBs:
    * HLYAOI: latest 2025-11-14 22:32Z, 10 rows returned.
    * MEAOI:  latest 2026-07-21 00:02Z, 10 rows returned.
  (HLYAOI post-reflow has been receiving no new inspection data
  since Nov 2025 - documented in memories/repo/aoi-db-freshness.md.)

+++ src/Nieweb.DataSources/Queries.cs
@@ -51,10 +51,10 @@ public sealed record TestedObjectQuery : BaseQuery
 }
 
 /// <summary>Keyset paging cursor for PANELS (ordered by Panel_Numeric_Date, Panel_Id).</summary>
-public readonly record struct PanelCursor(long LastPanelNumericDate, long LastPanelId);
+public readonly record struct PanelCursor(int LastPanelNumericDate, int LastPanelId);
 
 /// <summary>Keyset paging cursor for CARDS (ordered by Panel_Id, Card_Id_On_Panel).</summary>
-public readonly record struct CardCursor(long LastPanelId, int LastCardIdOnPanel);
+public readonly record struct CardCursor(int LastPanelId, int LastCardIdOnPanel);
 
 /// <summary>Keyset paging cursor for TESTED_OBJECT (ordered by Panel_Id, Card_Id_On_Panel, Object_Id).</summary>
-public readonly record struct TestedObjectCursor(long LastPanelId, int LastCardIdOnPanel, int LastObjectId);
+public readonly record struct TestedObjectCursor(int LastPanelId, int LastCardIdOnPanel, int LastObjectId);
